IEN-345 - rename component file
